centralize boxplot computing function
