+++ client/plots/barchart.data.js
@@ -2,6 +2,7 @@ import Partjson from 'partjson'
 import { compute_bins } from '#shared/termdb.bins.js'
 import { sample_match_termvaluesetting } from '../common/termutils'
 import { isNumeric } from '#shared/helpers.js'
+import { boxplot_getvalue } from '#shared/boxplot.js'
 
 export function getBarchartData(_q, data) {
 	/*
@@ -166,11 +167,7 @@ function getCharts(q, data) {
 				const values = context.self.values.filter(d => d !== null && !isNaN(d))
 				if (!values.length) return
 				values.sort((i, j) => i - j)
-				const stat = boxplot_getvalue(
-					values.map(v => {
-						return { value: +v }
-					})
-				)
+				const stat = boxplot_getvalue(values)
 				stat.mean = context.self.sum / values.length
 				let s = 0
 				for (const v of values) {
@@ -308,37 +305,6 @@ function getNumericIdVal(d, term, q, rows) {
 
 function getUndefinedIdVal() {
 	return [['-'], undefined]
-}
-
-function boxplot_getvalue(lst) {
-	/* ascending order
-  each element: {value}
-  */
-	const l = lst.length
-	if (l < 5) {
-		// less than 5 items, won't make boxplot
-		return { out: lst }
-	}
-	const p50 = lst[Math.floor(l / 2)].value
-	const p25 = lst[Math.floor(l / 4)].value
-	const p75 = lst[Math.floor((l * 3) / 4)].value
-	const p05 = lst[Math.floor(l * 0.05)].value
-	const p95 = lst[Math.floor(l * 0.95)].value
-	const p01 = lst[Math.floor(l * 0.01)].value
-	const iqr = p75 - p25
-
-	let w1, w2
-	if (iqr == 0) {
-		w1 = 0
-		w2 = 0
-	} else {
-		const i = lst.findIndex(i => i.value > p25 - iqr * 1.5)
-		w1 = lst[i == -1 ? 0 : i].value
-		const j = lst.findIndex(i => i.value > p75 + iqr * 1.5)
-		w2 = lst[j == -1 ? l - 1 : Math.max(0, j - 1)].value
-	}
-	const out = lst.filter(i => i.value < p25 - iqr * 1.5 || i.value > p75 + iqr * 1.5)
-	return { w1, w2, p05, p25, p50, p75, p95, iqr, out }
 }
 
 export function getCategoryData(q, data) {
