+++ server/src/mds3.densityPlot.js
@@ -1,4 +1,4 @@
-import { getDensity } from '../routes/termdb.violinBox.ts'
+import { getDensity } from './routes/termdb.violinBox.ts'
 
 /*
 term: numeric term from termdb
